+++ .gitignore
@@ -1,3 +1,6 @@
 server/node_modules
+server/serverkeys
+server/clientkeys
 node-client/node_modules
-
+node-client/serverkeys
+node-client/clientkeys
