--- node-client/clientkeys/privkey.pem
@@ -1,15 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIICXQIBAAKBgQCgXABUnlVfiSMUOPq0xHery3F/cVJUeo8QlUnkmCipy6MvtN6g
-xT7sS1wGzLV9bSMeR2Yvw7mZrDxd6znKX5UvT6K+vvHYH9b+FUhuh5ooHsFtxNrd
-4Ff3BHF42aDNWjQQXxOP8gck5o3yZdkcsqd5uMm6dR6ajFES5LaoFQ2SiQIDAQAB
-AoGAPKUultHedaf2dUJrRgv7OiRRgyIHQd2r11iuPl5xGm3iBDvB6klOV6zUW9n1
-VT7+Bgh4k93QhlaF6fr4uxlU4JtUqIcMVJzOMBXSA3r4prQ4UuMcF82bGEhk8OrC
-RWMjh92/RQTlLGEaY60ipQS62lsAyK7kzeEwdj645oLn1YUCQQDOjOeouRk9xxiF
-Xi7f3VkOzH/wN0IzBvKnXOkPsUs5sydy3zUM8Qp6zo0w/AjZ/HR+d+CTLs19cjtf
-nmO5VZrLAkEAxsAiDnSymcXPJgXawMiY7iZ0Y96BBXwBvCpKUfmBToeXsaN+454e
-5pc9N9sKiUWMlAjUXCcX89YerwH29L85ewJBAIFlBDtOueFZuKLUPclAgWoj8oIq
-G+6bi6gNLHQIPXgII85yxmk1F7uXahwlx+XHobMYpjeOOm08ArYEI7UXS8ECQQCQ
-TWmss0AFc+COC8wjEsIujw8Mf72PnNSpqM7jK9Odj8csu7diTijDZWWWS1wgBLYn
-uS1uZxf1+y22BSgn1TtHAkAXuCKzDEwonetDDzdRjFjSEF/96D0cd9PkdDX3wCDB
-iLWaa3U1p79R0bkiAvTHJT2JTQcgd3lxQI7rFMvO8Ab7
------END RSA PRIVATE KEY-----
