--- server/clientkeys/pubkey.pem
@@ -1,6 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQCgXABUnlVfiSMUOPq0xHery3F/
-cVJUeo8QlUnkmCipy6MvtN6gxT7sS1wGzLV9bSMeR2Yvw7mZrDxd6znKX5UvT6K+
-vvHYH9b+FUhuh5ooHsFtxNrd4Ff3BHF42aDNWjQQXxOP8gck5o3yZdkcsqd5uMm6
-dR6ajFES5LaoFQ2SiQIDAQAB
------END PUBLIC KEY-----
