--- server/serverkeys/privkey.pem
@@ -1,15 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIICXQIBAAKBgQDaBLIDNoHJsSYCAFl6f/Xi3QmDxQV78tspf5j8KaAz0oDgqvQx
-GzF7s3piOoNA9JcbDAvHXtstduL+wjTyO04E1jvANbQnUSK7B+sh9mqmS+EvFPo3
-FlQG+TbBNNabZ+5e1ecvxFiFOZJbeq7JBnt1Z6SYZsGnrOj/doE88YmqJQIDAQAB
-AoGAVeaPtRBe5vczgIpetxN/8FkjDJYuaP3xOFKcFo7YHlMMkHkLKeYLAzBeQiHH
-4Wn9t5xxb6aCOVGkcrhWYDLZs/fJ9r4vd1u/RCh2xSffaBH8a8vJ+4Eb/oNI9Wxw
-k24khgW3NMRt//5o4xj5sXR+w7fS29MVv67O7YsdbzKTFKkCQQD3+8Q+irG+Ejax
-vNRTbNPP1EVo9QxP3WtWDWe0WLO39VqmH/dFevxDXlnOrW6LkzZzKMT94a39BNU2
-XTeCsjIjAkEA4RDyYxV/N/zylVLzBB+hLaAodx8Ckq1qnBaQv7Qi081uwpKr6xRL
-v5uX2dcEKq5EPkcqfrrKRc4c3b6EsHhDFwJBAMkYAA7bd8FzkajWQLQaJzV8pIC1
-XrOREY6ptjQYTjzaGRVoDKPqVJDo/RjNqTxEN75xDDX5AZeRiLjaJL62dFMCQERV
-WuKOklOzbwu7ZaFo4QzzRZX1SliyYl4F0+kAjlp0Ux6sXZpATUddvFsdl1xGEN4/
-yT01BVjwXX9N6weRW/0CQQDA+58wAXFd+5FkH7A7I24OA2sFs6R+ysK1uQ3u7145
-rzvBTRRxWorImGehQmCwSfP9TaomBUuvt+GvkFdo29Hs
------END RSA PRIVATE KEY-----
