--- server/serverkeys/pubkey.pem
@@ -1,6 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDaBLIDNoHJsSYCAFl6f/Xi3QmD
-xQV78tspf5j8KaAz0oDgqvQxGzF7s3piOoNA9JcbDAvHXtstduL+wjTyO04E1jvA
-NbQnUSK7B+sh9mqmS+EvFPo3FlQG+TbBNNabZ+5e1ecvxFiFOZJbeq7JBnt1Z6SY
-ZsGnrOj/doE88YmqJQIDAQAB
------END PUBLIC KEY-----
